rename resize.py to resizing.py
