+++ pystiche/image/transforms/__init__.py
@@ -1,6 +1,6 @@
 from .core import *
 from .io import *
-from .resize import *
+from .resizing import *
 from .motif import *
 from .color import *
 from .misc import *
