+++ pystiche/image/transforms/resizing.py
